requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
cycler==0.11.0
docopt==0.6.2
fonttools==4.38.0
idna==3.6
importlib-metadata==6.7.0
kiwisolver==1.4.5
matplotlib==3.5.3
MouseInfo==0.1.3
numpy==1.21.6
opencv-python==4.9.0.80
packaging==24.0
Pillow==9.5.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Quartz==9.2
pyparsing==3.1.2
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
python-dateutil==2.9.0.post0
pytweening==1.2.0
requests==2.31.0
rubicon-objc==0.4.6
six==1.16.0
typing-extensions==4.7.1
urllib3==1.26.15
yarg==0.1.9
zipp==3.15.0
